sched/debug: Make sd->flags sysctl writable again
The flags of sched domain in sysctl was limited to read-only in a
recent change. However this might bring inconvenience to some performance
tuning/function test which are sensitive to load balance. Since
the user might want to evaluate the impact of different combination
of sched domain flags on the workloads.

Per Commit 9818427c6270 ("sched/debug: Make sd->flags sysctl read-only"),
we can make the flags writable if we update the cached pointers after
the flag of sched domain has been modified.

Leverage update_domain_cpu() to update corresponding cached pointer of that
CPU once the flags has been changed.

link: https://lore.kernel.org/all/a18a80656e67bb527fdd25e8cd4d355b2bfaa310.1594062828.git.yu.c.chen@intel.com/
Signed-off-by: Chen Yu <yu.c.chen@intel.com>
Signed-off-by: guojinhui.liam <guojinhui.liam@bytedance.com>