nbd: do not set sndtimeo if WAIT_ON_TIMEOUT is set
If NBD_RT_WAIT_ON_TIMEOUT is set, it means the user does not want
the socket to be dropped by the kernel. We already do it in nbd_xmit_timeout.
We missed it here. So if the nbd server hangs and the user prefers to
wait, do not set sndtimeo here.

Signed-off-by: Hou Pu <houpu@bytedance.com>

(Not upstream yet)