x86/fault: Add a helper function to sanitize error code
commit cd072dab453a9b4a9f7927f9eddca5a156fbd87d upstream.

vDSO exception fixup is a replacement for signals in limited situations.
Signals and vDSO exception fixup need to provide similar information to
userspace, including the hardware error code.

That hardware error code needs to be sanitized.  For instance, if userspace
accesses a kernel address, the error code could indicate to userspace
whether the address had a Present=1 PTE.  That can leak information about
the kernel layout to userspace, which is bad.

The existing signal code does this sanitization, but fairly late in the
signal process.  The vDSO exception code runs before the sanitization
happens.

Move error code sanitization out of the signal code and into a helper.
Call the helper in the signal code.

Intel-SIG: commit cd072dab453a x86/fault: Add a helper function to
sanitize error code.
Backport summary: This is a dependency patch for CET preparation patch

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Jarkko Sakkinen <jarkko@kernel.org>
Signed-off-by: Borislav Petkov <bp@suse.de>
Acked-by: Jethro Beekman <jethro@fortanix.com>
Link: https://lkml.kernel.org/r/20201112220135.165028-18-jarkko@kernel.org
[ Jing Han amend commit log ]
Signed-off-by: Jing Han <jing1.han@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>