PCI/ERR: Clear AER status only when we control AER
commit aa344bc8b727b47b4350b59d8166216a3f351e55 upstream
Backport summary: Backport to kernel 5.10.18 to enable PCIe/RCEC.

In some cases a bridge may not exist as the hardware controlling may be
handled only by firmware and so is not visible to the OS. This scenario is
also possible in future use cases involving non-native use of RCECs by
firmware. In this scenario, we expect the platform to retain control of the
bridge and to clear error status itself.

Clear error status only when the OS has native control of AER.

Signed-off-by: Sean V Kelley <sean.v.kelley@intel.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Youquan Song <youquan.song@intel.com>
Signed-off-by: Fam Zheng <fam.zheng@bytedance.com>