MAINTAINERS: Fix the tree location for INTEL SGX patches
commit 3ac517313b929619dbb7ceae005ec66d0859b23b upstream.

After a discussion with Boris et al, I've come to realize that a
disjoint GIT tree for SGX does not make any sense.

Instead, follow the pattern of other MAINTAINERS entries, IRQ DOMAINS for
instance, and re-define T-entry so that it will reference the pre-existing
topic branch for SGX. As Boris explained to me, reviewed patches will be
routinely picked to this branch.

Fixes: bc4bac2ecef0 ("x86/sgx: Update MAINTAINERS")
Signed-off-by: Jarkko Sakkinen <jarkko@kernel.org>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210121024256.54565-1-jarkko@kernel.org
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>