KVM: x86/xen: Fix __user pointer handling for hypercall page installation
commit 92f4d400a407235783afd4399fa26c4c665024b5 upstream.

The address we give to memdup_user() isn't correctly tagged as __user.
This is harmless enough as it's a one-off use and we're doing exactly
the right thing, but fix it anyway to shut the checker up. Otherwise
it'll whine when the (now legacy) code gets moved around in a later
patch.

Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>