KVM: vmx/pmu: Add PMU_CAP_LBR_FMT check when guest LBR is enabled
commit 9c9520ce883386dc3794c7d60204487ff1db09cb upstream.

Usespace could set the bits [0, 5] of the IA32_PERF_CAPABILITIES
MSR which tells about the record format stored in the LBR records.

The LBR will be enabled on the guest if host perf supports LBR
(checked via x86_perf_get_lbr()) and the vcpu model is compatible
with the host one.

Signed-off-by: Like Xu <like.xu@linux.intel.com>
Message-Id: <20210201051039.255478-4-like.xu@linux.intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>