mfd: intel_pmt: Fix nuisance messages and handling of disabled capabilities
commit a1a5c1c3df282dc122508a17500317266ef19e46 upstream.

Backport summary: backport to kernel 5.10.18 for intel PMT (Platform
Monitoring Technology) support

Some products will be available that have PMT capabilities that are not
supported. Remove the warnings in this instance to avoid nuisance messages
and confusion.

Also return an error code for capabilities that are disabled by quirk to
prevent them from keeping the driver loaded if only disabled capabilities
are found.

Fixes: 4f8217d5b0ca ("mfd: Intel Platform Monitoring Technology support")
Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: yingbao jia<yingbao.jia@intel.com>
Signed-off-by: Fam Zheng <fam.zheng@bytedance.com>