x86/pkeys: Move read_pkru() and write_pkru()
commit 784a46618f634973a17535b7d3d03cd4ebc0ccbd upstream.

write_pkru() was originally used just to write to the PKRU register.  It
was mercifully short and sweet and was not out of place in pgtable.h with
some other pkey-related code.

But, later work included a requirement to also modify the task XSAVE
buffer when updating the register.  This really is more related to the
XSAVE architecture than to paging.

Move the read/write_pkru() to asm/pkru.h.  pgtable.h won't miss them.

Intel-SIG: commit 784a46618f63 x86/pkeys: Move read_pkru() and write_pkru().
Intel Advanced Matrix Extensions(AMX) is one of the key features on Sapphire
Rapids(SPR) and it depends on xsave, this patch is part of Thomas Gleixner's
x86 FPU code cleanup that upstream AMX implementation is building on, and is
also key to all code for being developed in upstream that depends on xsave.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210623121455.102647114@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>