ACPI: Do not singal PRM support if not enabled
commit 392ed6a789803fbfd49994e95fe99cd07b07eb87 upstream.

If the OS confirms PRM (Platform Runtime Mechanism) support through
the \_SB._OSC PRM bit, the BIOS may start relying on the presence of
PRM support in the OS, so prevent the PRM bit from being set in the
\_SB._OSC capabilities bitmask when PRM support is not built in so
as to avoid confusing the BIOS in that case.

Intel-SIG: commit 392ed6a78980 ACPI: Do not singal PRM support if
not enabled.
Platform Runtime Mechanism feature backport.

Fixes: 60faa8f1ac6e ("ACPI: Add \_SB._OSC bit for PRM")
Signed-off-by: Aubrey Li <aubrey.li@linux.intel.com>
[ rjw: Rewrite subject and changelog, replace #ifdef with if (IS_ENABLED()) ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
[ Aubrey Li: amend commit log ]
Signed-off-by: Aubrey Li <aubrey.li@linux.intel.com>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>