veth: factor out initialization helper
commit dedd53c5e075e4277f2a01c6d78318ac16e645c3 upstream.

differences with up.. code:
there is no logical difference.We only adapted the code
based on the existing implementation.

Extract in simpler helpers the code to enable and disable a
range of xdp/napi instance, with the common property that
"disable" helpers can't fail.

Will be used by the next patch. No functional change intended.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Albert Huang <huangjie.albert@bytedance.com>