dmaengine: idxd: clear block on fault flag when clear wq
commit bd2f4ae5e019efcfadd6b491204fd60adf14f4a3 upstream.

The block on fault flag is not cleared when we disable or reset wq. This
causes it to remain set if the user does not clear it on the next
configuration load. Add clear of flag in dxd_wq_disable_cleanup()
routine.

Intel-SIG: commit bd2f4ae5e019 dmaengine: idxd: clear block on fault flag when clear wq.
Incremental backporting patches for DSA/IAX on Intel Xeon platform.

Fixes: da32b28c95a7 ("dmaengine: idxd: cleanup workqueue config after disabling")
Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/162803023553.3086015.8158952172068868803.stgit@djiang5-desk3.ch.intel.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>
[ Xiaochen Shen: amend commit log ]
Signed-off-by: Xiaochen Shen <xiaochen.shen@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>