x86/mce: Drop copyin special case for #MC
commit 690658471b5f28d306e6492c4585d748cb5304e8 upstream.

Fixes to the iterator code to handle faults that are not on page
boundaries mean that the special case for machine check during copy from
user is no longer needed.

For a full list of those fixes, see the output of:

  git log --oneline v5.14 ^v5.13 -- lib/iov_iter.c

Intel-SIG: commit 690658471b5f x86/mce: Drop copyin special case for #MC.
backport for MCA recovery enhance and bug fix.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210818002942.1607544-4-tony.luck@intel.com
[ Youquan Song amend commit log ]
Signed-off-by: Youquan Song <youquan.song@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>