x86/fpu: Do not leak fpstate pointer on fork
commit f0cbc8b3cdf7d1c724155cd9cecffe329bb96119 upstream.

If fork fails early then the copied task struct would carry the fpstate
pointer of the parent task.

Not a problem right now, but later when dynamically allocated buffers
are available, keeping the pointer might result in freeing the
parent's buffer. Set it to NULL which prevents that. If fork reaches
clone_thread(), the pointer will be correctly set to the new task
context.

Intel-SIG: commit f0cbc8b3cdf7 x86/fpu: Do not leak fpstate pointer on
fork.
The second part of the effort to support AMX: Move register state into
a container struct.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20211013145322.817101108@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>