x86/fpu/signal: Use fpstate for size and features
commit 5509cc78080d29b23706dbf076d51691b69f3c79 upstream.

For dynamically enabled features it's required to get the features which
are enabled for that context when restoring from sigframe.

The same applies for all signal frame size calculations.

Intel-SIG: commit 5509cc78080d x86/fpu/signal: Use fpstate for size and
features.
The second part of the effort to support AMX: Move register state into
a container struct.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/87ilxz5iew.ffs@tglx
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>