bytedance: config: enable CONFIG_PRINTK_CALLER
This option is intended for environments where multiple threads
concurrently call printk() for many times, for it is difficult to
interpret without knowing where these lines (or sometimes individual
line which was divided into multiple lines due to race) came from.
Selecting this option causes printk() to add a caller "thread id"
(if in task context) or a caller "processor id" (if not in task
context) to every message.

Signed-off-by: Muchun Song <songmuchun@bytedance.com>