selftests: Provide local define of __cpuid_count()
commit a23039c7306f53416ba35d230201398ea34f4640 upstream.

Duplication of the CPUID wrappers should be avoided.

Both gcc and clang/LLVM provide __cpuid_count() macros but neither
the macro nor its header file are available in all the compiler
versions that need to be supported by the selftests. __cpuid_count()
as provided by gcc is available starting with gcc v4.4, so it is
not available if the latest tests need to be run in all the
environments required to support kernels v4.9 and v4.14 that
have the minimal required gcc v3.2.

Duplicate gcc's __cpuid_count() macro to provide a centrally defined
macro for __cpuid_count() to help eliminate the duplicate CPUID wrappers
while continuing to compile in older environments.

Intel-SIG: commit a23039c7306f selftests: Provide local define of __cpuid_count().

Suggested-by: Shuah Khan <skhan@linuxfoundation.org>
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>
Signed-off-by: Shuah Khan <skhan@linuxfoundation.org>
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Yunhui Cui <cuiyunhui@bytedance.com>