perf/x86/intel/uncore: Add Emerald Rapids
commit 5268a2842066c227e6ccd94bac562f1e1000244f upstream.

From the perspective of the uncore PMU, the new Emerald Rapids is the
same as the Sapphire Rapids. The only difference is the event list,
which will be supported in the perf tool later.

Intel-SIG: commit 5268a2842066 perf/x86/intel/uncore: Add Emerald Rapids
Backport to 5.10 for EMR core and uncore PMU support.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lore.kernel.org/r/20230106160449.3566477-4-kan.liang@linux.intel.com
[ Yunying Sun: amend commit log ]
Signed-off-by: Yunying Sun <yunying.sun@intel.com>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>